+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.0.12 - 2022.10.04
+### Added
+* Add `ObjectAutocomplete` class to allow for easily adding all of the properties of an object as autocomplete items
+* Add missing Twig tags `else`, `elseif`, `endblock` & `endif`
+* Allow the `twigfieldOptions` config object to be passed into the Twig macros
+* Include a hash of the `twigfieldOptions` in the cache key used for the autocomplete
+
+### Changed
+* Refactor to `ObjectParserAutocomplete` & `ObjectParserInterface`
+
 ## 1.0.11 - 2022.08.24
 ### Changed
 * Remove `FluentModel` class and replace the magic method setter with fluent setter methods in `CompleteItem`